Module arduboy_rust::prelude::arduboy2

This is the Module to interact in a save way with the Arduboy2 C++ library.

All of the functions are safe wrapped inside the Arduboy2 struct.

Structs

Arduboy2

This is the struct to interact in a save way with the Arduboy2 C++ library.

Point

This struct is used by a few Arduboy functions.

Rect

This struct is used by a few Arduboy functions.

Enums

Color

This item is to chose between Black or White

Constants

FONT_SIZE

The standard font size of the arduboy

HEIGHT

The standard height of the arduboy

WIDTH

The standard width of the arduboy